requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
azure-ai-agents==1.0.0
azure-ai-projects==1.0.0b11
azure-common==1.1.28
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.19.0
azure-mgmt-core==1.6.0
azure-mgmt-resourcegraph==8.0.0
azure-monitor-opentelemetry==1.6.9
azure-monitor-opentelemetry-exporter==1.0.0b40
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
Deprecated==1.2.18
distlib==0.4.0
fastapi==0.115.13
filelock==3.19.1
fixedint==0.1.6
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
identify==2.6.13
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
nodeenv==1.9.1
oauthlib==3.3.1
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==25.0
platformdirs==4.4.0
pre_commit==4.3.0
propcache==0.3.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
ruff==0.12.11
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
